Move averaging over batch from loss to trainer
This is necessary because some loss function are unware of the batch size and
may average differently, which causes problems when the loss is computed "in
parts" (on different parts of the batch and summed). All loss functions now
report the sum of the losses, averaging over the batch is done in the trainer.

Change of scale in loss functions: bce, mr
As before: kl, ce